Restructure reducers, use separate files
