+++ redux/reducers/login.js
@@ -1,0 +1,4 @@
+import { LOGIN_REQUEST, LOGIN_SUCCESS, LOGIN_FAILURE } from '../actions'
+import { LOGOUT } from '../actions'
+
+
