+++ redux/reducers/token.js
@@ -1,0 +1,19 @@
+import { LOGIN_SUCCESS } from '../actions'
+import { LOGOUT } from '../actions'
+
+
+/**
+ * Authentication token with the Dakara server
+ */
+
+function token(state = null, action) {
+    if (action.type === LOGIN_SUCCESS) {
+        return action.payload.token
+    } else if (action.type === LOGOUT) {
+        return null
+    } else {
+        return state
+    }
+}
+
+export default token
